Restrict page access and hide edit and delete if user is not the host

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic, User
@@ -7,6 +9,10 @@ from .forms import RoomForm
 
 
 def login_page(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -62,6 +68,7 @@ def room(request, room_id):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='/login')
 def create_room(request):
     form = RoomForm()
 
@@ -75,9 +82,13 @@ def create_room(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def update_room(request, room_id):
     room = Room.objects.get(id=room_id)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You're not allowed here!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -90,8 +101,12 @@ def update_room(request, room_id):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def delete_room(request, room_id):
     room = Room.objects.get(id=room_id)
+
+    if request.user != room.host:
+        return HttpResponse("You're not allowed here!")
 
     if request.method == 'POST':
         room.delete()
